Add PHPUnit tests for Prof M and the associated infrastructure
Works locally and with Jenkins. Note that Download from Jive is
currently broken, so Travis builds have not been tested recently.

+++ scripts/GetCopyOfSugar.sh
@@ -1,15 +1,10 @@
 #!/usr/bin/env bash
 
-# This script downloads a Sugar dev build from the SugarCRM Developer Builds community.
+# This script gets a copy of Sugar from the designated directory or downloads a Sugar dev build from the SugarCRM
+# Developer Builds community.
 #
 # Note: you must have access to the SugarCRM Developer Builds community
-# (https://community.sugarcrm.com/community/developer/developer-builds) in order for the script to work.
-#
-# Run the script with the following arguments:
-# 1: Email address associated with your SugarCRM Developer Builds Community account
-# 2: Password associated with the above account
-# 3: Sugar version to download (Example: 7.11)
-# 4: Sugar Edition to download (Options: Ult, Ent, Pro)
+# (https://community.sugarcrm.com/community/developer/developer-builds) in order for the download to be successful.
 
 
 ######################################################################
@@ -21,11 +16,14 @@ then
     echo "Not all required command line arguments were set. Please run the script again with the required arguments:
         1: Email address associated with your SugarCRM Developer Builds Community account
         2: Password associated with the above account
-        3: Sugar version to download (Example: 7.11)
-        4: Sugar edition to download (Options: Ult, Ent, Pro)
+        3: Sugar name (For example: SugarEnt-7.11)
+        4. The path to where the Sugar download should be stored
+        5. (Optional) Path to where Sugar source zip files are stored. If this param is not provided, the Sugar
+           source zip files will be downloaded from the SugarCRM Developer Builds Community.  The Sugar source zip files
+           should be named with the following pattern: Sugar$sugarEdition-$sugarVersion. For example: SugarEnt-7.11
 
-        For example: ./downloadSugarFromCommunity.sh email@example.com mypassword 7.11 Pro"
-    exit
+        For example: ./GetCopyOfSugar.sh email@example.com mypassword SugarEnt-7.11 ../sugar_source_zips"
+    exit 1
 fi
 
 # Email address associated with your SugarCRM developer community account
@@ -34,11 +32,14 @@ email=$1
 # Password associated with your SugarCRM developer community account
 password=$2
 
-# The Sugar version to download
-sugarVersion=$3
+# The Sugar name (For example: SugarEnt-7.11)
+sugarName=$3
 
-# The Sugar edition to download
-sugarEdition=$4
+# The path to where the Sugar download should be stored
+sugarDirectory=$4
+
+# Path to where existing Sugar source zip files are stored
+sugarSourceZipsDirectory=$5
 
 # The name of the cookie jar file where the cookies required for this script will be stored
 cookieFile="mycookie"
@@ -63,11 +64,11 @@ checkStatusCode(){
         else
             echo "Status code is not the expected $1: $statusCode"
             echo "$2"
-            exit
+            exit 1
         fi
     else
         echo "Unable to find status code in response: $2"
-        exit
+        exit 1
     fi
 }
 
@@ -82,7 +83,7 @@ getLocationFromResponse(){
         echo "$location"
     else
         echo "Unable to find location in response"
-        exit
+        exit 1
     fi
 }
 
@@ -95,19 +96,50 @@ getHiddenFormFieldValue(){
     if [[ $response =~ $regexToken ]]
     then
         value="${BASH_REMATCH[1]}"
-        echo "$value"
+
+        # This is a hack specifically for Travis CI.  Travis CI outputs a string like the following in the middle
+        # of the SAML Response:
+        #
+        # 100 12223    0 12223    0     0  16084      0 --:--:-- --:--:-- --:--:-- 16104
+        # * Connection #0 to host auth.sugarcrm.com left intact
+        #
+        # This Regex Token pulls this junk out of the SAML Response
+        newLineRegexToken="(.*)[[:space:]]+[[:digit:]][[:digit:]][[:digit:]][[:space:]].*intact[[:space:]]*(.*)"
+        if [[ $1 == 'SAMLResponse' && $value =~ $newLineRegexToken ]]
+        then
+            value="${BASH_REMATCH[1]}${BASH_REMATCH[2]}"
+        fi
+
+        echo $value
     else
         echo "Unable to find $2 in response"
         echo "$2"
-        exit
+        exit 1
     fi
 
 }
 
 
 ######################################################################
-# Setup
+# Check if we need to download the Sugar source zip
 ######################################################################
+
+# If we already have a copy of the Sugar source zip, we'll copy it to the Sugar directory and exit the script
+if [[ -d "$sugarSourceZipsDirectory" && -e "$sugarSourceZipsDirectory/$sugarName.zip" ]]
+then
+    echo "$sugarSourceZipsDirectory/$sugarName.zip already exists. A new copy of Sugar will not be downloaded."
+
+    cp $sugarSourceZipsDirectory/$sugarName.zip $sugarDirectory/$sugarName.zip
+    exit 0
+fi
+
+
+######################################################################
+# Setup for download
+######################################################################
+
+# Change to the Sugar directory
+cd $sugarDirectory
 
 # Delete the cookie jar file if it exists
 rm -f $cookieFile
@@ -117,6 +149,7 @@ rm -f $cookieFile
 # Authenticate to the community
 ######################################################################
 
+#TODO: Replace ./mycookie with variable
 response="$(curl -v -c ./mycookie -b ./mycookie 'https://community.sugarcrm.com/login.jspa?ssologin=true&fragment=&referer=%2Fcommunity%2Fdeveloper%2Fdeveloper-builds' 2>&1)"
 checkStatusCode "302" "$response"
 location="$(getLocationFromResponse "$response")"
@@ -140,13 +173,13 @@ response="$(curl -v -c ./mycookie -b ./mycookie $location 2>&1)"
 checkStatusCode "200" "$response"
 samlResponse="$(getHiddenFormFieldValue "SAMLResponse" "$response")"
 
-response="$(curl -v -c ./mycookie -b ./mycookie --data-urlencode "SAMLResponse=$samlResponse&RelayState=L2NvbW11bml0eS9kZXZlbG9wZXIvZGV2ZWxvcGVyLWJ1aWxkcw%3D%3D" 'https://community.sugarcrm.com/saml/sso' 2>&1)"
+response="$(curl -v -c ./mycookie -b ./mycookie --data-urlencode "SAMLResponse=$samlResponse&RelayState=L2NvbW11bml0eS9kZXZlbG9wZXIvZGV2ZWxvcGVyLWJ1aWxkcw==" 'https://community.sugarcrm.com/saml/sso' 2>&1)"
 checkStatusCode "302" "$response"
 
 
-######################################################################
-# Download the Sugar zip
-######################################################################
+#######################################################################
+## Download the Sugar zip
+#######################################################################
 
 sugarVersion_7_10="7.10"
 sugarVersion_7_11="7.11"
@@ -156,41 +189,41 @@ sugarEdition_Ent="Ent"
 sugarEdition_Pro="Pro"
 
 # Get the url for the appropriate Sugar version and edition
-if [[ "$sugarVersion$sugarEdition" == "$sugarVersion_7_10$sugarEdition_Ult" ]]
+if [[ "$sugarName" == "Sugar$sugarEdition_Ult-$sugarVersion_7_10" ]]
 then downloadUrl="https://community.sugarcrm.com/servlet/JiveServlet/downloadBody/5839-102-1-8005/SugarUlt-7.10.2.0-dev.1.zip"
 
-elif [[ "$sugarVersion$sugarEdition" == "$sugarVersion_7_10$sugarEdition_Ent" ]]
+elif [[ "$sugarName" == "Sugar$sugarEdition_Ent-$sugarVersion_7_10" ]]
 then downloadUrl="https://community.sugarcrm.com/servlet/JiveServlet/downloadBody/5837-102-1-8003/SugarEnt-7.10.2.0-dev.1.zip"
 
-elif [[ "$sugarVersion$sugarEdition" == "$sugarVersion_7_10$sugarEdition_Pro" ]]
+elif [[ "$sugarName" == "Sugar$sugarEdition_Pro-$sugarVersion_7_10" ]]
 then downloadUrl="https://community.sugarcrm.com/servlet/JiveServlet/downloadBody/5838-102-1-8004/SugarPro-7.10.2.0-dev.1.zip"
 
-elif [[ "$sugarVersion$sugarEdition" == "$sugarVersion_7_11$sugarEdition_Ult" ]]
+elif [[ "$sugarName" == "Sugar$sugarEdition_Ult-$sugarVersion_7_11" ]]
 then downloadUrl="https://community.sugarcrm.com/servlet/JiveServlet/downloadBody/5958-102-1-8147/SugarUlt-7.11.0.0-dev.1.zip"
 
-elif [[ "$sugarVersion$sugarEdition" == "$sugarVersion_7_11$sugarEdition_Ent" ]]
+elif [[ "$sugarName" == "Sugar$sugarEdition_Ent-$sugarVersion_7_11" ]]
 then downloadUrl="https://community.sugarcrm.com/servlet/JiveServlet/downloadBody/5959-102-1-8148/SugarEnt-7.11.0.0-dev.1.zip"
 
-elif [[ "$sugarVersion$sugarEdition" == "$sugarVersion_7_11$sugarEdition_Pro" ]]
+elif [[ "$sugarName" == "Sugar$sugarEdition_Pro-$sugarVersion_7_11" ]]
 then downloadUrl="https://community.sugarcrm.com/servlet/JiveServlet/downloadBody/5957-102-1-8146/SugarPro-7.11.0.0-dev.1.zip"
 
 else
-    echo "Unable to find Sugar download URL for version $sugarVersion and edition $sugarEdition"
-    exit
+    echo "Unable to find Sugar download URL for $sugarName"
+    exit 1
 fi
 
 # Download the file
-fileName="Sugar$sugarEdition-$sugarVersion.zip"
-echo "Beginning download of $fileName from $downloadUrl"
-response="$(curl -v -c ./mycookie -b ./mycookie -o $fileName $downloadUrl 2>&1)"
+echo "Beginning download of $sugarName from $downloadUrl"
+response="$(curl -v -c ./mycookie -b ./mycookie -o $sugarName.zip $downloadUrl 2>&1)"
 checkStatusCode "200" "$response"
 echo "Download complete"
 
 # Check we didn't get an empty zip file
-fileSize=$(wc -c <"$fileName")
+fileSize=$(wc -c <"$sugarName.zip")
 if [[ $fileSize -lt 60000000 ]]
 then
-    echo "$fileName has a file size of $fileSize.  The download may not have been successful."
+    echo "$sugarName.zip has a file size of $fileSize.  The download may not have been successful."
+    exit 1
 fi
